+++ frontend/src/pages/index.ts
@@ -1,6 +1,8 @@
 export { AuthPage } from './AuthPage';
 export { DashboardPage } from './DashboardPage';
 export { ProjectPage } from './ProjectPage';
+export { CreateProjectPage } from './CreateProjectPage';
+export { EditProjectPage } from './EditProjectPage';
 export { ExportPage } from './ExportPage';
 export { ImportPage } from './ImportPage';